pyproject.toml
[project]
name = "mp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imageio>=2.37.3",
    "imageio-ffmpeg>=0.6.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.9",
    "mutagen>=1.47.0",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "soundfile>=0.13.1",
    "streamlit>=1.57.0",
]
